refactor(ai-chat): 整理
